git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@9382 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/dso-l2/src/com/tc/objectserver/persistence/sleepycat/ManagedObjectPersistorImpl.java
@@ -29,6 +29,8 @@ import com.tc.text.PrettyPrintable;
 import com.tc.text.PrettyPrinter;
 import com.tc.util.Assert;
 import com.tc.util.Conversion;
+import com.tc.util.NullSyncObjectIdSet;
+import com.tc.util.ObjectIDSet;
 import com.tc.util.SyncObjectIdSet;
 import com.tc.util.SyncObjectIdSetImpl;
 import com.tc.util.concurrent.ThreadUtil;
@@ -77,6 +79,11 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
                                                                          .getProperties()
                                                                          .getBoolean(
                                                                                      TCPropertiesConsts.L2_OBJECTMANAGER_PERSISTOR_LOGGING_ENABLED);
+  private static final boolean                 measurePerf           = TCPropertiesImpl
+                                                                         .getProperties()
+                                                                         .getBoolean(
+                                                                                     TCPropertiesConsts.L2_OBJECTMANAGER_PERSISTOR_MEASURE_PERF,
+                                                                                     false);
 
   private final Database                       objectDB;
   private final SerializationAdapterFactory    saf;
@@ -91,17 +98,22 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
   private final SleepycatCollectionsPersistor  collectionsPersistor;
   private final ObjectIDManager                objectIDManager;
   private final ConcurrentHashMap              statsRecords          = new ConcurrentHashMap();
+  private int                                  deleteCounter;
+  private int                                  deletePersistentStateCounter;
+  private long                                 deleteTime;
+  private final SyncObjectIdSet                extantObjectIDs;
+  private final SyncObjectIdSet                extantMapTypeOidSet;
 
   public ManagedObjectPersistorImpl(TCLogger logger, ClassCatalog classCatalog,
-                                    SerializationAdapterFactory serializationAdapterFactory, Database objectDB,
-                                    Database oidDB, Database oidLogDB, Database oidLogSeqDB,
-                                    CursorConfig dBCursorConfig, MutableSequence objectIDSequence, Database rootDB,
-                                    CursorConfig rootDBCursorConfig, PersistenceTransactionProvider ptp,
-                                    SleepycatCollectionsPersistor collectionsPersistor, boolean paranoid) {
+                                    SerializationAdapterFactory serializationAdapterFactory, DBEnvironment env,
+                                    MutableSequence objectIDSequence, Database rootDB, CursorConfig rootDBCursorConfig,
+                                    PersistenceTransactionProvider ptp,
+                                    SleepycatCollectionsPersistor collectionsPersistor, boolean paranoid)
+      throws TCDatabaseException {
     this.logger = logger;
     this.classCatalog = classCatalog;
     this.saf = serializationAdapterFactory;
-    this.objectDB = objectDB;
+    this.objectDB = env.getObjectDatabase();
     this.objectIDSequence = objectIDSequence;
     this.rootDB = rootDB;
     this.rootDBCursorConfig = rootDBCursorConfig;
@@ -114,14 +126,50 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
       this.objectIDManager = new NullObjectIDManager();
     } else if (oidFastLoad) {
       // read objectIDs from compressed DB
-      MutableSequence sequence = new SleepycatSequence(this.ptp, logger, 1, 1000, oidLogSeqDB);
-      this.objectIDManager = new FastObjectIDManagerImpl(oidDB, oidLogDB, ptp, dBCursorConfig, sequence);
+      MutableSequence sequence = new SleepycatSequence(this.ptp, logger, 1, 1000, env.getOidLogSequeneceDB());
+      this.objectIDManager = new FastObjectIDManagerImpl(env, ptp, sequence, this);
     } else {
       // read objectIDs from object DB
-      this.objectIDManager = new PlainObjectIDManagerImpl(objectDB, ptp, dBCursorConfig);
+      this.objectIDManager = new PlainObjectIDManagerImpl(objectDB, ptp);
     }
 
+    this.extantObjectIDs = getAllObjectIDs();
+    this.extantMapTypeOidSet = getAllMapsObjectIDs();
+
     if (STATS_LOGGING_ENABLED) startStatsPrinter();
+    if (measurePerf) startDeleteOperPrinter();
+  }
+
+  public int getObjectCount() {
+    return extantObjectIDs.size();
+  }
+
+  public boolean addNewObject(ObjectID id) {
+    return extantObjectIDs.add(id);
+  }
+
+  public boolean containsObject(ObjectID id) {
+    return extantObjectIDs.contains(id);
+  }
+
+  public void removeAllObjectsByID(SortedSet<ObjectID> ids) {
+    this.extantObjectIDs.removeAll(ids);
+  }
+
+  public ObjectIDSet snapshotObjects() {
+    return this.extantObjectIDs.snapshot();
+  }
+
+  public boolean containsMapType(ObjectID id) {
+    return extantMapTypeOidSet.contains(id);
+  }
+
+  public boolean addMapTypeObject(ObjectID id) {
+    return extantMapTypeOidSet.add(id);
+  }
+
+  public void removeAllMapTypeObject(Collection ids) {
+    extantMapTypeOidSet.removeAll(ids);
   }
 
   public long nextObjectIDBatch(int batchSize) {
@@ -193,6 +241,19 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
     SyncObjectIdSet rv = new SyncObjectIdSetImpl();
     rv.startPopulating();
     Thread t = new Thread(objectIDManager.getObjectIDReader(rv), "ObjectIdReaderThread");
+    t.setDaemon(true);
+    t.start();
+    return rv;
+  }
+
+  public SyncObjectIdSet getAllMapsObjectIDs() {
+    SyncObjectIdSet rv = new SyncObjectIdSetImpl();
+
+    Runnable reader = objectIDManager.getMapsObjectIDReader(rv);
+    if (reader == null) return new NullSyncObjectIdSet();
+
+    rv.startPopulating();
+    Thread t = new Thread(reader, "MapsObjectIdReaderThread");
     t.setDaemon(true);
     t.start();
     return rv;
@@ -278,7 +339,7 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
     try {
       status = basicSaveObject(persistenceTransaction, managedObject);
       if (OperationStatus.SUCCESS.equals(status) && managedObject.isNew()) {
-        status = objectIDManager.put(persistenceTransaction, managedObject.getID());
+        status = objectIDManager.put(persistenceTransaction, managedObject);
       }
     } catch (DBException e) {
       throw e;
@@ -352,6 +413,20 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
     t.start();
   }
 
+  private void startDeleteOperPrinter() {
+    Thread t = new Thread(new Runnable() {
+      public void run() {
+        while (true) {
+          ThreadUtil.reallySleep(60000);
+          logger.info("Deletes count:" + deleteCounter + " delete state count:" + deletePersistentStateCounter
+                      + " usedTime(ns): " + deleteTime);
+        }
+      }
+    }, "Delete Statistics printer");
+    t.setDaemon(true);
+    t.start();
+  }
+
   private static String createLogString(String className, long written, int count, int newCount) {
     StringBuilder sb = new StringBuilder();
     appendFixedSpaceString(sb, className, 50);
@@ -413,11 +488,13 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
 
         // record new object-IDs to be written to persistent store later.
         if (managedObject.isNew()) {
-          objectIDManager.prePutAll(oidSet, managedObject.getID());
+          objectIDManager.prePutAll(oidSet, managedObject);
         }
       }
-      if (!OperationStatus.SUCCESS.equals(objectIDManager.putAll(persistenceTransaction, oidSet))) { throw new DBException(
-                                                                                                                           "Failed to save Object-IDs"); }
+      if (!OperationStatus.SUCCESS.equals(objectIDManager.putAll(persistenceTransaction, oidSet))) {
+        //
+        throw new DBException("Failed to save Object-IDs");
+      }
     } catch (Throwable t) {
       throw new DBException(t);
     }
@@ -455,7 +532,17 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
         // make the formatter happy
         throw new DBException("Unable to remove ManagedObject for object id: " + id + ", status: " + status);
       } else {
-        collectionsPersistor.deleteCollection(tx, id);
+        long startTime = 0;
+        if (measurePerf) {
+          startTime = System.nanoTime();
+          ++deleteCounter;
+        }
+        if (containsMapType(id)) {
+          if (measurePerf) ++deletePersistentStateCounter;
+          // may return false if ManagedObject persistent state empty
+          collectionsPersistor.deleteCollection(tx, id);
+        }
+        if (measurePerf) deleteTime += (System.nanoTime() - startTime);
       }
     } catch (DatabaseException t) {
       throw new DBException(t);
@@ -525,6 +612,8 @@ public final class ManagedObjectPersistorImpl extends SleepycatPersistorBase imp
     out.println(this.getClass().getName());
     out = out.duplicateAndIndent();
     out.println("db: " + objectDB);
+    out.indent().print("extantObjectIDs: ").visit(extantObjectIDs).println();
+    out.indent().print("extantMapTypeOidSet: ").visit(extantMapTypeOidSet).println();
     return out;
   }
